requirements.txt
flask
flask_bootstrap
pymongo
pymysql
werkzeug
pyserial